refactor tests to direct the majority towards scan_and_tag_s3_object itself
